% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_ensemble_functions.R
\name{ar_ensemble}
\alias{ar_ensemble}
\title{AR Ensemble}
\usage{
ar_ensemble(ens, obs_col, mem_col, train = 90)
}
\arguments{
\item{ens}{A data frame with one observation column, at least
one forecast column, and at least one additional column (e.g. date).}

\item{obs_col}{The observation column.}

\item{mem_col}{The column(s) of the forecast members(s).}

\item{train}{The length of the rolling training period used for
fitting an autoregressive process.}
}
\value{
A list with four elements:
\describe{
\item{\code{observation}}{A numeric vector containing the
observations for the forecast period (original period except for the
first \code{train} dates.}
\item{\code{forecast}}{A data frame containing the AR modified forecasts.}
\item{\code{variance}}{A data frame containing variance estimates
corresponding to the forecasts.}
\item{\code{additional}}{A data frame containg the variables (columns) of
\code{ens} which were
not specified by \code{obs_col} and \code{mem_col}.}
}
}
\description{
For each member of a forecast ensemble,
a corresponding autoregressive modified member is generated.
}
\note{
It is assumed that in each row of the data frame \code{ens}
the forecast matches the observation,
i.e. for each row the difference between forecast and observation
is the forecast error.
}
\examples{
#
ar_ensemble(Magdeburg[1:(90+1),], obs_col = 6, mem_col = 7:56)
}
\author{
J. Gross, A. Moeller.
}

