% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_slp_functions.R
\name{combineSLP}
\alias{combineSLP}
\title{Spread-Adjusted Linear Pool (SLP) of two Normals}
\usage{
combineSLP(x, pred_one = list(mean = NULL, sd = NULL), pred_two = pred_one,
  train = 90, weight_grid = seq(0, 1, 0.1), scale_grid = seq(0.6, 1.4,
  0.1))
}
\arguments{
\item{x}{A vector of observations of a weather quantity.}

\item{pred_one}{A list with two vectors \code{mean} and \code{sd} of
the same length as \code{x} respectively, providing parameters of the
normal predictive distribution.}

\item{pred_two}{A list with two vectors \code{mean} and \code{sd}
of the same length as \code{x} respectively, providing parameters of the second
normal predictive distribution.}

\item{train}{The length of the training period.}

\item{weight_grid}{The possible values of the SLP weight.}

\item{scale_grid}{The possible values of the SLP scale parameter.}
}
\value{
A data frame with the observation, predictive mean and
standard deviation of the two normals, and SLP weight and
scale parameter for the forecast period.
}
\description{
Computes the weight and scale parameters of the
SLP of two normals from a rolling training period.
}
\details{
For each forecast date (all dates except for the first
\code{train} dates), the optimal combination of SLP weight and scale
parameters is found on a grid determined by \code{weight_grid} and
\code{scale_grid}, such that the corresponding
(predictive distribution, observation)-pair minimizes the average
CRPS with respect to the rolling training period of
length \code{train}.
}
\examples{
combineSLP(16.1, list(mean = 15, sd = 0.8), list(mean = 18, sd = 1), 0)


}
\author{
J. Gross, A. Moeller.
}
\references{
Gneiting T.,  Ranjan R. 2013. Combining predictive
distributions. \emph{Electronic Journal of Statistics},
\strong{7},  1747--1782.
}

