% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_utility_functions.R
\name{convert_chardate}
\alias{convert_chardate}
\title{Conversion of Character Vector of Dates}
\usage{
convert_chardate(x)
}
\arguments{
\item{x}{A character vector with dates in format 'yyyymmdd' with no NA's and no duplicates.}
}
\value{
A list with two elements:
 \describe{
\item{\code{dates}}{Date object in format yyyy-mm-dd, chronologically ordered.}
\item{\code{missing}}{Non-occuring dates between chronolocically first and last input date.}
}
}
\description{
Converts a character vector into a Date object
and checks for missing dates.
}
\note{
If \code{missing} is 'None' and no warning is given, the
dates in \code{x} are complete and in chronological order.

The function is not thoroughly protected against an incorrect input format.
}
\examples{
x <- c('20020105', '20020102', '20020103')
convert_chardate(x)

}
\author{
J. Gross, A. Moeller.
}

