% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_verification.R
\name{crps_norm}
\alias{crps_norm}
\title{CRPS for a Normal or a Mixture of two Normals}
\usage{
crps_norm(x, mu1, sd1, mu2 = mu1, sd2 = sd1, w1 = 1)
}
\arguments{
\item{x}{A vector of observations for which the CRPS is to be computed.}

\item{mu1}{A vector of expectations of the first normal. Must be of the same length as \code{x}.}

\item{sd1}{A vector of standard deviations of the first normal. Must be of the same length as \code{x}.}

\item{mu2}{A vector of expectations of the second normal. Must be of the same length as \code{x}.}

\item{sd2}{A vector of standard deviations of the second normal. Must be of the same length as \code{x}.}

\item{w1}{A vector of weights between 0 and 1 associated with the first normal.
Must be either of length one, or of the same length as \code{x}.}
}
\value{
A vector of CRPS values.
}
\description{
Computes the continuous ranked probability score
(CRPS) of given observations when the predictive distribution is
the normal distribution or a mixture of two normals.
}
\details{
Formula (5) from Grimit et al. (2006) is applied
for the special case of two normals.
}
\examples{
crps_norm(0.5, -1, 1, 2, 1, 0.2)
}
\author{
J. Gross, A. Moeller.
}
\references{
Grimit E.P., Gneiting T., Berrocal V., Johnson N.A.
2006. The continuous ranked probability score for circular
variables and its application to mesoscale forecast ensemble
verification. \emph{Quarterly Journal of the Royal Meteorological
Society}, \strong{132}, 2925--2942.
}

