% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_statistic_functions.R
\name{cx_comb}
\alias{cx_comb}
\title{Convex Combination}
\usage{
cx_comb(x, y, w)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector of the same length as x.}

\item{w}{A weight vector of the same length as x.}
}
\value{
A vector of the same length as x and y.
}
\description{
Computes a convex combination of two vector of the same length
}
\author{
J. Gross, A. Moeller.
}

