% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_statistic_functions.R
\name{dm_test}
\alias{dm_test}
\title{Diebold-Mariano Test}
\usage{
dm_test(s1, s2, alternative = c("two.sided", "less", "greater"), h = 1)
}
\arguments{
\item{s1}{Score from method 1.}

\item{s2}{Score from method 2.}

\item{alternative}{A character string specifying the alternative
hypothesis, must be one of 'two.sided' (default), 'greater' or
'less'.}

\item{h}{An integer specifying the smallest lag not
used in the autocovariance function.}
}
\value{
A object of class \code{'htest'}.
}
\description{
Performs the Diebold-Mariano test for
equal predictive performance of two methods with respect to
a scoring rule.
}
\details{
The null hypothesis is that the difference \code{s1 - s2} has zero mean.
The alternative \code{'less'} is that \code{s1 - s2} has negative mean.

The test statistic requires values of the autocovariance function
of  \code{s1 - s2} with lags smaller than \code{h}, i.e.
\code{lag.max = h - 1} (the truncation lag)
in \code{\link[stats]{acf}}.

 The difference \code{s1 - s2}  may contain missing values,
 in which case complete cases are used and a warning is given.
}
\note{
The function \code{dm_test} is inspired by the function \code{dm.test} from
package \pkg{forecast}. It is a 'simpler' version in the sense that it computes
p-values from the normal instead of the t distribution.
}
\examples{
set.seed(1)
s1 <- arima.sim(list(ar = 0.7), sd = 0.5, 100)
s2 <- arima.sim(list(ar = 0.7), sd = 0.5, 100) - 0.1
dm_test(s1, s2)
}
\references{
Diebold F.X, Mariano R.S. 1995. Comparing predictive accuracy.
\emph{Journal of Business & Economic Statistics}, \strong{13}, 253--263.

Gneiting T., Katzfuss  M. 2014. Probabilistic forecasting.
\emph{Annual Review of Statistics and Its Application},
\strong{1}, 125--151.
}
\seealso{
\code{\link[forecast]{dm.test}} in package \pkg{forecast}.
}
\author{
J. Gross, A. Moeller.
}
