% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_aux_functions.R
\name{.cx_comb}
\alias{.cx_comb}
\title{Convex Combination}
\usage{
.cx_comb(x, y, w)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector of the same length as x.}

\item{w}{A weight vector of the same length as x.}
}
\value{
A vector of the same length as x and y.
}
\description{
Computes a convex combination of two vector of the same length
}
\examples{
\dontrun{
x <- c(1, 2, 3)
y <- c(3, 2, 1)
w <- c(1/2, 1/3, 1/2)
.cx_comb(x, y, w)
}

}
\author{
J. Gross, A. Moeller.
}
