% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_aux_functions.R
\name{.is_weight}
\alias{.is_weight}
\title{Check for Weight}
\usage{
.is_weight(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical vector of length one.
}
\description{
Checks whether all elements of vector lie between 0 and 1.
}
\examples{
\dontrun{
a <- c(0.3, 0.4, 0.9)
.is_weight(a)
}

}
\author{
J. Gross, A. Moeller.
}
