% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_ensemble_functions.R
\name{ensembleAR}
\alias{ensembleAR}
\title{AR Ensemble and Predictive Distribution}
\usage{
ensembleAR(ens, obs_col, mem_col, skip = 0, train_ar = 90,
  train_crps = 30)
}
\arguments{
\item{ens}{A data frame with one observation column, at least
one forecast column, and at least one additional column (e.g. date).}

\item{obs_col}{The observation column.}

\item{mem_col}{The column(s) of the forecast members(s).}

\item{skip}{A number corresponding to the forecast ahead time (0 for ahead times
not greater than 24 hours, 1 for ahead times greater than 24 hours and not greater
than 48 hours, and so on).}

\item{train_ar}{The length of the rolling training period used for
fitting an autoregressive process.}

\item{train_crps}{The length of the additional training period used
for computing the predictive standard deviation.}
}
\value{
A data frame.
}
\description{
A wrapper for sequential processing of the functions
\code{\link{ar_ensemble}} and \code{\link{ar_preddistr}}.
}
\examples{
ensembleAR(ens = Magdeburg[1:(90 + 30 + 1), -c(57,58)],
    obs_col = 6, mem_col = 7:56)

}
\author{
J. Gross, A. Moeller.
}
