% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_data.R
\docType{data}
\name{Magdeburg}
\alias{Magdeburg}
\title{Temperature Observation and Ensemble Forecasts}
\format{A data frame with 4461 rows and 58 variables:
\itemize{
\item{date.}{ Date in format 'yyyymmdd'}
\item{lat.}{ Latitude}
\item{lon.}{ Longitude}
\item{stationID.}{ Station identification number}
\item{stationName.}{ Station name}
\item{obs.}{ Observed temperature in degree Celsius}
\item{forc.1, forc.2, ..., forc.50.}{ Forecasts of 50 exchangeable ensemble members}
\item{forc.HRES.}{ High-resolution forecast}
\item{forc.CTRL.}{ Control forecast}
}}
\usage{
Magdeburg
}
\description{
The data set contains 24-h ahead forecasts for 2-m surface
temperature along with the verifying observations at station
'Magdeburg' for about 12 years.
}
\references{
Hemri S, Scheuerer M, Pappenberger F, Bogner K, Haiden T. 2014. Trends in the
predictive performance of raw ensemble weather forecasts. \emph{Geophysical
Research Letters} \strong{41}, 9197--9205, \doi{10.1002/2014GL062472}
}
\keyword{datasets}
