% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_utility_functions.R
\name{na_gaps}
\alias{na_gaps}
\title{Finding Gaps of NAs}
\usage{
na_gaps(x)
}
\arguments{
\item{x}{A vector or data frame.}
}
\value{
A list with three elements of class 'na_gaps'.
\describe{
\item{\code{table}}{A table of gap lengths and corresponding
frequencies.}
\item{\code{lengths}}{A vector of gap lengths, actually the names
of list element \code{table}.}
\item{\code{number}}{The number of incomplete cases.}
}
}
\description{
Computes lengths and frequencies of NA gaps, i.e.
sequences of consecutive NAs, in a vector or data frame.
}
\details{
If the input is a data frame, a missing value is always defined as a
not complete case, i.e. each row containing an NA is treated as a
missing observation.
}
\examples{
na_gaps(c(1, NA, NA, 1, 1, NA, NA, NA, 1, NA, NA, 1, 1, 1, NA, NA))
}
\seealso{
\code{\link[stats]{complete.cases}}
}
\author{
J. Gross, A. Moeller.
}
