% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_utility_functions.R
\name{previous_equal}
\alias{previous_equal}
\title{Relative Frequency of Previous Equal Elements}
\usage{
previous_equal(x, previous = 10)
}
\arguments{
\item{x}{A numeric vector.}

\item{previous}{The number of previous elements compared to the actual element.}
}
\value{
Relative frequency of possible successes as decribed in the Details section.
}
\description{
Compares each feasible element of a vector to a given number of
previous elements and computes the relative frequency of successes,
where a success occurs in case of positive equality comparison.
}
\details{
For each element of \code{x} greater than \code{previous},
a success is recorded when it is equal to the last \code{previous} elements,
where equality is assessed by \code{\link{all.equal}}.
}
\examples{
x <- c(1, 1, 2, 1, 2, 2, 2, 3)
previous_equal(x, previous = 1) # 3/7
previous_equal(x, previous = 2) # 1/6

}
\author{
J. Gross, A. Moeller.
}

