% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_utility_functions.R
\name{print.na_gaps}
\alias{print.na_gaps}
\title{Print NA Gaps}
\usage{
\method{print}{na_gaps}(x, ...)
}
\arguments{
\item{x}{An object of class 'na_gaps'}

\item{...}{Further arguments passed to S3 method for class 'table'}
}
\description{
Print method for objects of class 'na_gaps'.
}
\author{
J. Gross, A. Moeller.
}
\seealso{
\code{\link{na_gaps}}
}

