% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_ensemble_functions.R
\name{rank_ensemble}
\alias{rank_ensemble}
\title{Rank of Observation}
\usage{
rank_ensemble(ens, obs_col, mem_col)
}
\arguments{
\item{ens}{A data frame with one observation column and at least
one forecast column.}

\item{obs_col}{The observation column.}

\item{mem_col}{The column(s) of the forecast members(s).}
}
\value{
A vector of the same length as rows of \code{ens}.
}
\description{
Computes the rank of an observation
among the corresponding ensemble forecasts.
}
\details{
Ranks are computed with \code{\link[base]{rank}} and
\code{ties.method = 'first'}.
}
\examples{
x <- rank_ensemble(Magdeburg, obs_col = 6, mem_col = 7:56)
table(x)
}
\author{
J. Gross, A. Moeller.
}
