% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_statistic_functions.R
\name{var_ar}
\alias{var_ar}
\title{Variance of an AR process}
\usage{
var_ar(ar = numeric(), i_var = 1)
}
\arguments{
\item{ar}{A vector of autoregression coefficients.}

\item{i_var}{The innovations variance.}
}
\value{
The variance of the process.
}
\description{
Comptutes the variance of an autoregressive (AR) process.
}
\details{
The variance of an AR process of order \eqn{p} is
\deqn{\gamma_{0} = \frac{\sigma^2}{1 - \phi_{1} \rho_{1} - \phi_{2} \rho_{2} - \cdots
 - \phi_{p} \rho_{p}}}
where \eqn{\sigma^2} is the variance of the innovations and \eqn{\phi_{j}}
are the autoregression coefficients. The autocorrelations
\eqn{\rho_{j}} are computed by \code{\link{ARMAacf}}.
}
\examples{
## x <- arima.sim(list(ar = 0.7), sd = 0.5, 100000)
## var(x)
var_ar(ar = 0.7, i_var= 0.5^2)
}
\author{
J. Gross, A. Moeller.
}

