% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_verification.R
\name{veri_stats}
\alias{veri_stats}
\title{Verification Statistics}
\usage{
veri_stats(x, par_one = list(mean = NULL, sd = NULL), par_two = par_one,
  weight_one = NULL, scale = NULL)
}
\arguments{
\item{x}{A vector of observations of a weather quantity.}

\item{par_one}{A list with two vectors \code{mean} and \code{sd} of
the same length as \code{x} respectively, providing parameters of the
normal predictive distribution.}

\item{par_two}{A list with two vectors \code{mean} and \code{sd}
of the same length as \code{x} respectively, providing parameters of the second
normal predictive distribution. Only required if \code{weight_one}
is not \code{NULL} in which case the two parameter vectors must
be of the same length as \code{x}.}

\item{weight_one}{A vector of the same length as \code{x} providing
the weights (corresponding to the first normal) used for the spread-adjusted linear pool.}

\item{scale}{A vector of the same length as \code{x} providing the common scale
parameter used for the spread-adjusted linear pool.}
}
\value{
A list with elements
\describe{
\item{\code{pit}}{A vector of the same length as \code{x}
giving probability integral transform (PIT) values.}
\item{\code{crps}}{A vector of the same length as \code{x} giving
the continuous ranked probability score (CRPS) values.}
\item{\code{dss}}{A vector of the same length as \code{x} giving
the Dawid and Sebastiani score (DSS) values.}
\item{\code{pitvar}}{The sample variance of the PIT values.}
\item{\code{rmv}}{The square root of the mean of the predictive variance.}
}
}
\description{
Computes different verification statistics for
one predictive normal distribution, or for the spread-adjusted
linear pool (SLP) of two predictive normal distributions.
}
\details{
If \code{weight_one} is \code{NULL} verification statistics
are computed for a normal distribution with parameters provided by
\code{par_one}. Otherwise the predictive distribution is the
spread-adjusted linear pool (SLP) of two normals with parameters
provided by \code{par_one} and \code{par_two} and
SLP parameters \code{weight_one} (weights
corresponding to \code{par_one}) and common scale
parameter \code{scale}.
}
\examples{
veri_stats(17.5, par_one = list(mean = 15, sd = 1))
}
\author{
J. Gross, A. Moeller.
}

