% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_verification.R
\name{veri_synop}
\alias{veri_synop}
\title{Synopsis of Verification Statistics}
\usage{
veri_synop(x, method_name = NULL, digits = 4)
}
\arguments{
\item{x}{An object of class \code{'veri_stats'}.}

\item{method_name}{The name of the applied forecast method.}

\item{digits}{An integer specifying the digits for \code{round}. If
set to \code{NULL}, the values are not rounded.}
}
\value{
A data frame with one row and corresponding row name given by \code{method_name}.
}
\description{
Computes a synopsis for an object of
class 'veri_stats'.
}
\examples{
x <- veri_stats(17.5, pred_one = list(mean = 15,sd = 1.5))
veri_synop(x, method = "No method", digits = 3)
#' @author J. Gross, A. Moeller.
}

