% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_data.R
\docType{data}
\name{Magdeburg}
\alias{Magdeburg}
\title{Temperature Observation and Ensemble Forecasts}
\format{A data frame with 4461 rows and 58 variables:
\itemize{
\item{date.}{ Date in format 'yyyymmdd'}
\item{lat.}{ Latitude}
\item{lon.}{ Longitude}
\item{stationID.}{ Station identification number}
\item{stationName.}{ Station name}
\item{obs.}{ Observed temperature in degree Celsius}
\item{forc.1, forc.2, forc.3, forc.4, forc.5, forc.6, forc.7,
forc.8, forc.9, forc.10, forc.11, forc.12, forc.13, forc.14,
forc.15, forc.16, forc.17, forc.18, forc.19, forc.20, forc.21,
forc.22, forc.23, forc.24, forc.25, forc.26, forc.27, forc.28,
forc.29, forc.30, forc.31, forc.32, forc.33, forc.34, forc.35,
forc.36, forc.37, forc.38, forc.39, forc.40, forc.41, forc.42,
forc.43, forc.44, forc.45, forc.46, forc.47, forc.48, forc.49,
forc.50.}{ Forecasts of 50 exchangeable ensemble members}
\item{forc.HRES.}{ High-resolution forecast}
\item{forc.CTRL.}{ Control forecast}
}}
\usage{
Magdeburg
}
\description{
The data set contains 24-h ahead forecasts for 2-m surface
temperature along with the verifying observations at station
'Magdeburg' for about 12 years.
}
\references{
Hemri S, Scheuerer M, Pappenberger F, Bogner K, Haiden T. 2014. Trends in the
predictive performance of raw ensemble weather forecasts. \emph{Geophysical
Research Letters} \strong{41}, 9197--9205, \doi{10.1002/2014GL062472}
}
\keyword{datasets}

