% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_slp_functions.R
\name{slp_moments}
\alias{slp_moments}
\title{Predictive Moments of the SLP of two Normals}
\usage{
slp_moments(pred_one = list(mean = NULL, sd = NULL), pred_two = pred_one,
  weight_one = rep(0.5, length(pred_one$mean)), scale = rep(1,
  length(pred_one$sd)))
}
\arguments{
\item{pred_one}{A list with two vectors \code{mean} and \code{sd} of
the same length as \code{x} respectively, providing parameters of the
normal predictive distribution.}

\item{pred_two}{A list with two vectors \code{mean} and \code{sd}
of the same length as \code{x} respectively, providing parameters of the second
normal predictive distribution.}

\item{weight_one}{A vector of weights corresponding to \code{pred_one}.}

\item{scale}{A vector of scale values.}
}
\value{
A list with two elements \code{mu} (predictive mean) and
\code{sd} (predictive standard deviation) of the SLP combination.
}
\description{
Computes the predictive mean and predictive standard
deviation of the SLP of two normals.
}
\author{
J. Gross, A. Moeller.
}

