% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens_slp_functions.R
\name{slp_moments}
\alias{slp_moments}
\title{Moments of the SLP of two Normals}
\usage{
slp_moments(normal_one = list(mean = NULL, sd = NULL),
  normal_two = normal_one, weight_one = rep(0.5, length(normal_one$mean)),
  scale = rep(1, length(normal_one$sd)))
}
\arguments{
\item{normal_one}{A list with two vectors \code{mean} and \code{sd} of
the same length as \code{x} respectively, providing parameters of the
normal distribution.}

\item{normal_two}{A list with two vectors \code{mean} and \code{sd}
of the same length as \code{x} respectively, providing parameters of the second
normal distribution.}

\item{weight_one}{A vector of weights corresponding to \code{normal_one}.}

\item{scale}{A vector of scale values.}
}
\value{
A list with two elements \code{mean} and
\code{sd} of the SLP combination.
}
\description{
Computes the mean and standard
deviation of the SLP (spread-adjust linear pool) of two normals.
}
\examples{
slp_moments(list(mean = 15, sd = 0.8), list(mean = 18, sd = 1), 0.7, 1.2)
}
\author{
J. Gross, A. Moeller.
}

